fix(infra): 统一数据库到应用库 ecboot——等级列放宽000037/测试DSN单一事实源/时区双侧锁UTC
起因: 勘察发现三份事实源并行——测试硬编码 myuser@13306/mydatabase(容器),
而应用配置、make migrate-up、gf gen dao 全指向 root@3306/ecboot(本机 MySQL),
导致迁移打 A 库、测试跑 B 库、生成物取自 C 库(同一张表两侧列集都不同)。

- 迁移: 应用库补齐 35/36/37(原停 34); 新增 000037_user_level_bigint 把 user.level
  由 TINYINT 放宽为 BIGINT UNSIGNED, 对齐其列注释(存 user_level_rule.id, BIGINT)——
  原类型使规则 id>127 时 LevelRecalc 报 1264, 会员成长/升级链路失败
- 生成物: gf gen dao 重生成, 仅 entity/user.go 真变更(Level int→uint64); 其余 15 个
  文件是 CLI 换行符噪音已还原
- 测试 DSN: 7 处收敛为单一事实源 internal/testutil/db.go(ECBOOT_TEST_DSN 可覆盖)
- 时区(同时修掉一个真实缺陷): DSN 钉 loc=UTC&time_zone=UTC 双侧一致; 并给本机
  MySQL 载入时区表(原表在但 0 行→命名时区报 1298)。原缺陷: 库 time_zone=SYSTEM
  为 +08 而进程按 UTC 写墙钟, SQL 端 NOW() 快 8h——实测"开始时间=+60分钟"的轮播
  被 <NOW() 判为已开始; 推理 CancelTimeout 会把待付款单几乎立即全部取消。
  写法约束已写入注释: GoFrame 把 DSN 参数里的 + 变空格, 偏移量写法必失败——这也
  解释了批次 03 当年 time_zone='+08:00' 其实从未生效
- 工具链/文档同步: Makefile migrate-fresh 改到应用库同实例; compose 库名
  mydatabase→ecboot; AGENTS.md/两处 README 更新(顺带更正不存在的 manifest/config 路径)

验证: ecboot 70 表与 ecboot_fresh 全量重放(37 版本, 干净)一致;
go test ./... 连续两次全绿

+++ apps/server/internal/model/entity/user.go
@@ -21,7 +21,7 @@ type User struct {
 	ShareCode       string      `json:"shareCode"       orm:"share_code"       ` // 个人推广码(唯一;NULL=未生成,唯一索引不参与)
 	Gender          int         `json:"gender"          orm:"gender"           ` // 性别:0未知 1男 2女
 	GrowthValue     uint        `json:"growthValue"     orm:"growth_value"     ` // 成长值(只增不减,累计列)
-	Level           int         `json:"level"           orm:"level"            ` // 会员等级(user_level_rule.id,未启体系为NULL)
+	Level           uint64      `json:"level"           orm:"level"            ` // 会员等级(user_level_rule.id,未启体系为NULL)
 	LastLoginAt     *gtime.Time `json:"lastLoginAt"     orm:"last_login_at"    ` // 最后登录时间(休眠分级依据;登录链路一次命中,免login_log聚合)
 	LastActiveAt    *gtime.Time `json:"lastActiveAt"    orm:"last_active_at"   ` // 最后活跃时间(登录/下单/浏览任一;休眠判定以此为准)
 	Status          int         `json:"status"          orm:"status"           ` // 账号状态:1正常 2禁用
